+++ client/src/components/liberationmap/LiberationMap.tsx
@@ -1,4 +1,4 @@
-import { selectMapCenter } from "../../api/mapSlice";
+import { selectMapCenter, selectMapZoom } from "../../api/mapSlice";
 import { useAppSelector } from "../../app/hooks";
 import MapLayersControl from "../maplayers/MapLayersControl";
 import MapSearch from "../mapsearch";
@@ -11,8 +11,9 @@ import { MapContainer, ScaleControl } from "react-leaflet";
 export default function LiberationMap() {
   const map = useRef<Map>(null);
   const mapCenter = useAppSelector(selectMapCenter);
+  const wantedZoom = useAppSelector(selectMapZoom);
   useEffect(() => {
-    map.current?.setView(mapCenter, map.current?.getZoom() ?? 8, {
+    map.current?.setView(mapCenter, wantedZoom ?? map.current?.getZoom() ?? 8, {
       animate: true,
       duration: 1,
     });